@@ -6,18 +6,18 @@ import { check, group, sleep } from 'k6'
 import http from 'k6/http'
 
 //import { htmlReport } from "https://raw.githubusercontent.com/benc-uk/k6-reporter/main/dist/bundle.js";
-import { htmlReport } from '../dist/bundle.js'
-import { textSummary } from 'https://jslib.k6.io/k6-summary/0.0.1/index.js'
+import { htmlReport } from '../../dist/bundle.js'
+import { textSummary } from 'https://jslib.k6.io/k6-summary/0.1.0/index.js'
 
 const TARGET_URL = __ENV.TEST_TARGET || 'https://benc.dev'
-const RAMP_TIME = __ENV.RAMP_TIME || '1s'
-const RUN_TIME = __ENV.RUN_TIME || '2s'
+const RAMP_TIME = __ENV.RAMP_TIME || '5s'
+const RUN_TIME = __ENV.RUN_TIME || '20s'
 const USER_COUNT = __ENV.USER_COUNT || 20
 const SLEEP = __ENV.SLEEP || 0.5
 
 export function handleSummary(data) {
   return {
-    'summary.html': htmlReport(data, { debug: false }),
+    'summary.html': htmlReport(data, { title: 'Simple Load Test', debug: true }),
     stdout: textSummary(data, { indent: ' ', enableColors: true }),
   }
 }
@@ -27,6 +27,7 @@ export let options = {
     { duration: RAMP_TIME, target: USER_COUNT },
     { duration: RUN_TIME, target: USER_COUNT },
   ],
+  summaryTrendStats: ['avg', 'min', 'max', 'med', 'p(90)', 'p(95)', 'p(99)'],
   thresholds: {
     http_req_duration: ['p(95)<1000'],
     iteration_duration: ['max<4000'],
